refactor ファイル構造をblogと揃える
