+++ tests/combination/Button/normalButton.test.ts
@@ -1,4 +1,4 @@
-import Button from '$lib/components/Button/normalButton.svelte';
+import Button from '$lib/components/atom/button/NormalButton.svelte';
 import { expect, test } from '@playwright/experimental-ct-svelte';
 test.describe('ボタンの表示テスト', () => {
 	test('normalButton.svelte', async ({ mount }) => {
